Move mem_lock logic into mem.h and mem.cpp
